requirements.txt
flask
google-generativeai
openpyxl
pandas
python-dotenv
Pillow
